+++ config.py
@@ -1,5 +1,6 @@
 import os
 import connexion
+import yaml
 from flask_marshmallow import Marshmallow
 
 from gcpdac.local_logging import get_logger
@@ -29,3 +30,16 @@ celery_app = make_celery(__name__)
 
 def get_celery():
     return celery_app
+
+def read_config_map():
+    # Returns the EC configuration as a dictionary
+    try:
+        with open("/app/ec-config.yaml", 'r') as stream:
+            try:
+                return yaml.safe_load(stream)
+            except yaml.YAMLError as exc:
+                logger.exception("Failed to parse EC YAML after successfully opening - {}".format(exc))
+                raise
+    except Exception:
+        logger.exception("Failed to load EC YAML file")
+        raise
