+++ gcpdac/terraform_utils.py
@@ -1,5 +1,7 @@
 from python_terraform import Terraform
 
+import config
+
 logger = config.logger
 
 def terraform_init(backend_prefix, terraform_state_bucket, tf: Terraform):
